backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
pydantic[email]>=2.0.0
python-dotenv>=1.0.1
firebase-admin>=6.3.0
httpx>=0.27.0
pytest>=8.0.0
gunicorn
psycopg2-binary
